fix(backend): Исправить типы added_at с i64 на f64 для Specta совместимости
Проблема:
- Specta не поддерживает BigInt (i64) по умолчанию
- Генерация TypeScript биндингов падала с ошибкой BigIntForbidden

Решение:
- Изменить все поля added_at с i64 на f64 в resource структурах:
  - EffectResource
  - FilterResource
  - TransitionResource
  - TemplateResource
  - StyleTemplateResource
  - SubtitleResource
- Обновить все timestamp преобразования: Utc::now().timestamp() as f64

Файлы:
- src-tauri/src/state/project_state.rs: изменение типов полей
- src-tauri/src/state/commands/resources.rs: преобразование timestamp

+++ src-tauri/src/state/commands/resources.rs
@@ -61,7 +61,7 @@ impl ResourceCommands {
       name: name.clone(),
       effect_id,
       parameters,
-      added_at: Utc::now().timestamp(),
+      added_at: Utc::now().timestamp() as f64,
     };
 
     // Add to effects pool
@@ -131,7 +131,7 @@ impl ResourceCommands {
       name: name.clone(),
       filter_id,
       parameters,
-      added_at: Utc::now().timestamp(),
+      added_at: Utc::now().timestamp() as f64,
     };
 
     // Add to filters pool
@@ -210,7 +210,7 @@ impl ResourceCommands {
       name: name.clone(),
       transition_id,
       parameters,
-      added_at: Utc::now().timestamp(),
+      added_at: Utc::now().timestamp() as f64,
     };
 
     // Add to transitions pool
@@ -277,7 +277,7 @@ impl ResourceCommands {
       name: name.clone(),
       template_id,
       data: data.clone(),
-      added_at: Utc::now().timestamp(),
+      added_at: Utc::now().timestamp() as f64,
     };
 
     // Add to templates pool
@@ -353,7 +353,7 @@ impl ResourceCommands {
       name: name.clone(),
       template_id,
       data: data.clone(),
-      added_at: Utc::now().timestamp(),
+      added_at: Utc::now().timestamp() as f64,
     };
 
     // Add to style templates pool
@@ -420,7 +420,7 @@ impl ResourceCommands {
       name: name.clone(),
       style_id,
       data: data.clone(),
-      added_at: Utc::now().timestamp(),
+      added_at: Utc::now().timestamp() as f64,
     };
 
     // Add to subtitles pool
